Move centrifuge related cli commands to separate package
