+++ chains/evm/cli/cli.go
@@ -4,6 +4,7 @@ import (
 	"github.com/ChainSafe/chainbridge-core/chains/evm/cli/account"
 	"github.com/ChainSafe/chainbridge-core/chains/evm/cli/admin"
 	"github.com/ChainSafe/chainbridge-core/chains/evm/cli/bridge"
+	"github.com/ChainSafe/chainbridge-core/chains/evm/cli/centrifuge"
 	"github.com/ChainSafe/chainbridge-core/chains/evm/cli/deploy"
 	"github.com/ChainSafe/chainbridge-core/chains/evm/cli/erc20"
 	"github.com/ChainSafe/chainbridge-core/chains/evm/cli/erc721"
@@ -73,6 +74,10 @@ func init() {
 	// erc721
 	EvmRootCLI.AddCommand(erc721.ERC721Cmd)
 
+	// centrifuge
+	EvmRootCLI.AddCommand(centrifuge.CentrifugeCmd)
+
+	// account
 	EvmRootCLI.AddCommand(account.AccountRootCMD)
 
 	// utils
